requirements.txt
Django==1.10.5
djangorestframework==3.5.4
Markdown==2.6.8
pkg-resources==0.0.0
